+++ nominatim/server/falcon/server.py
@@ -2,7 +2,7 @@
 #
 # This file is part of Nominatim. (https://nominatim.org)
 #
-# Copyright (C) 2022 by the Nominatim developer community.
+# Copyright (C) 2023 by the Nominatim developer community.
 # For a full list of authors see the git log.
 """
 Server implementation using the falcon webserver framework.
@@ -13,8 +13,7 @@ from pathlib import Path
 import falcon
 import falcon.asgi
 
-from nominatim.api import NominatimAPIAsync
-from nominatim.apicmd.status import StatusResult
+from nominatim.api import NominatimAPIAsync, StatusResult
 import nominatim.result_formatter.v1 as formatting
 
 CONTENT_TYPE = {
